extract mention toolbar and integrate in interpretation window
