+++ src/ui/InterpretationWindow.js
@@ -1,4 +1,5 @@
 import { SharingWindow } from './SharingWindow';
+import { MentionToolbar } from './MentionToolbar.js';
 
 export var InterpretationWindow;
 
@@ -20,9 +21,11 @@ InterpretationWindow = function(c, sharing, interpretation, success) {
         emptyText: i18n.write_your_interpretation + '...',
         enableKeyEvents: true,
         value: interpretation ? interpretation.text : undefined,
+        mentionToolbar: MentionToolbar(c),
         listeners: {
-            keyup: function() {
+            keyup: function(f, e) {
                 shareButton.xable();
+                this.mentionToolbar.displayMentionSuggestion(f, e);
             }
         }
     });
@@ -128,8 +131,8 @@ InterpretationWindow = function(c, sharing, interpretation, success) {
         },
         listeners: {
             show: function(w) {
-                uiManager.setAnchorPosition(w, 'favoriteButton');
-
+                uiManager.setAnchorPosition(w, 'favoriteButton', {y: 130});
+                
                 uiManager.enableRightClick();
 
                 if (!w.hasDestroyOnBlurHandler) {
